refactor(app): separating app from server
