+++ package.json
@@ -56,7 +56,7 @@
     "uuid": "^8.3.2"
   },
   "scripts": {
-    "dev": "tsnd --inspect --transpile-only --ignore-watch node_modules --respawn -r tsconfig-paths/register src/shared/infra/http/app.ts",
+    "dev": "tsnd --inspect --transpile-only --ignore-watch node_modules --respawn -r tsconfig-paths/register src/shared/infra/http/server.ts",
     "typeorm": "tsnd -r tsconfig-paths/register ./node_modules/typeorm/cli",
     "test": "jest",
     "seed:admin": "tsnd src//shared/infra/typeorm/seed/admin.ts"
