+++ src/shared/infra/http/app.ts
@@ -1,5 +1,24 @@
-import { app } from "./server";
+import "reflect-metadata";
+import express from "express";
+import "express-async-errors";
+import swaggerUi from "swagger-ui-express";
 
-const PORT = 3333;
+import errorHandler from "@shared/infra/http/middlewares/errorHandler";
+import createConnection from "@shared/infra/typeorm";
+import "@shared/container";
 
-app.listen(PORT, () => console.log(`Server listen on port ${PORT}`));
+import swaggerFile from "../../../swagger.json";
+import { router } from "./routes";
+
+createConnection();
+const app = express();
+
+app.use(express.json());
+
+app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(swaggerFile));
+
+app.use(router);
+
+app.use(errorHandler);
+
+export { app };
